deskagent_agent/requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
dj-database-url==3.1.0
Django==5.2.10
django-allauth==65.13.1
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-environ==0.12.0
django-htmx==1.27.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
packaging==25.0
pillow==12.1.0
psutil==7.2.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
python-decouple==3.8
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
service-identity==24.2.0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
websockets==16.0
whitenoise==6.11.0
zope.interface==8.2
